requirements.txt
flask==3.1.1
werkzeug==3.1.3
psycopg2-binary==2.9.10
qrcode[pil]==8.1
gunicorn==23.0.0
